finish lib refactor
