+++ lib/download.js
@@ -9,7 +9,7 @@ const env = require('../env.json');
 const Authorization = `Bearer ${env.app_token}`;
 
 
-async function handleDownload (fileId, saveAsName) {
+module.exports = async function handleDownload (fileId, saveAsName) {
     // find req download file meta data before attempting to DL
     const metaResult = await getMetaData(fileId)
     .catch(e => {
