+++ lib/get-info.js
